--- src/main/java/server/nanum/dto/Response/MyReviewDto.java
@@ -1,13 +1,0 @@
-package server.nanum.dto.Response;
-
-import lombok.Builder;
-
-@Builder
-public record MyReviewDto(
-        Long orderId,
-        String orderName,
-        String imgUrl,
-        Float rating,
-        String comment){
-
-}
