--- src/main/java/server/nanum/dto/Response/MyUnReviewDto.java
@@ -1,10 +1,0 @@
-package server.nanum.dto.Response;
-
-import lombok.Builder;
-
-@Builder
-public record MyUnReviewDto(
-        Long orderId,
-        String orderName,
-        String imgUrl){
-}
