+++ src/main/java/server/nanum/dto/response/MyReviewDto.java
@@ -1,0 +1,16 @@
+package server.nanum.dto.response;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategy;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.Builder;
+
+@Builder
+@JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
+public record MyReviewDto(
+        Long orderId,
+        String orderName,
+        String imgUrl,
+        Float rating,
+        String comment){
+
+}
