+++ src/main/java/server/nanum/dto/response/MyUnReviewDto.java
@@ -1,0 +1,13 @@
+package server.nanum.dto.response;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategy;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.Builder;
+
+@Builder
+@JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
+public record MyUnReviewDto(
+        Long orderId,
+        String orderName,
+        String imgUrl){
+}
