+++ src/main/java/server/nanum/service/ReviewService.java
@@ -8,9 +8,9 @@ import server.nanum.domain.DeliveryStatus;
 import server.nanum.domain.Order;
 import server.nanum.domain.Review;
 import server.nanum.domain.User;
-import server.nanum.dto.Request.AddReviewDto;
-import server.nanum.dto.Response.MyReviewOrdersDto;
-import server.nanum.dto.Response.MyUnReviewOrdersDto;
+import server.nanum.dto.request.AddReviewDto;
+import server.nanum.dto.response.MyUnReviewOrdersDto;
+import server.nanum.dto.response.MyReviewOrdersDto;
 import server.nanum.repository.OrderRepository;
 import server.nanum.repository.ReviewRepository;
 import server.nanum.repository.UserRepository;
